web/requirements.txt
gunicorn
flask
flask-wtf
flask-login
flask-sqlalchemy
flask-migrate
flask-markdown
flask-moment
pymysql
celery
celery-redbeat
redis
requests
lxml
fuzzywuzzy
python-Levenshtein
flower